third_party/blink/web_tests/platform/generic/fast/forms/label/label-selection-by-dragging-expected.txt

Test the value of checkbox after selecting the associated label's text

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS checkbox.checked is true
PASS window.getSelection().toString() is "Some, Text associated with input"
PASS successfullyParsed is true

TEST COMPLETE
